test: move screenshot path validation test to tests directory
Addresses Greptile review feedback: embedded test was not discoverable
by pytest. Now properly located in tests/test_screenshot.py for CI/CD.
